Working with sizes. Needs testing on devices.

+++ push.sh
@@ -1,4 +1,8 @@
-rm -r bin
-git add *
-git commit -m "$1"
-git push notasgithub
+if [ $# -ne 1  ]; then
+	echo "Wrong number of arguments used. Usage: push.sh <commit message>."
+else
+	rm -r bin;
+	git add * &&
+	git commit -m "$1" &&
+	git push notasgithub
+fi
